requirements.txt
aiogram==3.22.0
python-dotenv==1.2.1
ollama==0.6.1
requests==2.32.5
sqlalchemy==2.0.45
asyncpg==0.31.0
playwright==1.57.0
fake-useragent==2.2.0
loguru==0.7.3
